refactor: mv color_copied_files to TableContext, so we can use it in binder.
